[#21] 슬라이더 뷰 크기 및 간격 조정
